add DTB compiling for vanilla kernel

+++ arm-soc_jessie/2_chroot_build-deb.sh
@@ -26,11 +26,14 @@ touch ../build_begin.txt
 export XZ_DEFAULTS=-7   # limit memory usage
 LOCAL_INST=../local_install
 make ARCH=arm CROSS_COMPILE=arm-linux-gnueabi- mrproper
-cp -a ../config .config
+[ -e ../config ] && cp -a ../config .config
 make ARCH=arm CROSS_COMPILE=arm-linux-gnueabi- oldconfig
+[ ! -e ../config ] && cp -a .config ../config
 make ARCH=arm CROSS_COMPILE=arm-linux-gnueabi- -j$PARALLEL zImage modules 2>&1 | tee log.0_zImage_modules
-make ARCH=arm CROSS_COMPILE=arm-linux-gnueabi- modules_install INSTALL_MOD_PATH=$LOCAL_INST
-mkdir -p $LOCAL_INST/boot
+make ARCH=arm CROSS_COMPILE=arm-linux-gnueabi- dtbs 2>&1 | tee log.1_zImage_modules
+make ARCH=arm CROSS_COMPILE=arm-linux-gnueabi- modules_install INSTALL_MOD_PATH=$LOCAL_INST 2>&1 | tee log.2_modules_install
+mkdir -p $LOCAL_INST/boot $LOCAL_INST/dtbs
 cp -a arch/arm/boot/zImage $LOCAL_INST/boot/vmlinuz-`cat include/config/kernel.release`
+cp -a arch/arm/boot/dts/*.dtb $LOCAL_INST/dtbs/
 
 touch ../build_end_binary.txt
